requirements.txt
Appium-Python-Client==2.9.0
async-generator==1.10
attrs==23.1.0
certifi==2022.12.7
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
outcome==1.2.0
PySocks==1.7.1
selenium==4.9.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.10.2
typing-extensions==4.5.0
urllib3==1.26.15
wsproto==1.2.0
zipp==3.15.0
